src/requirements.txt
docarray==0.40.0
langchain==0.2.5
langchain-cli==0.0.25
langchain-community==0.2.5
langchain-openai==0.1.10
openai==1.35.3
python-dotenv==1.0.1
tiktoken==0.7.0
yt-dlp==2024.5.27
